[task-9/feature]: add unsigned decimal strategy and int decimal strategy

+++ shared/utils.h
@@ -1,5 +1,5 @@
 char* s21_itoa(long int value, char* buffer, int base);
-void s21_ftoa(double n, char* res, int afterpoint);
+void s21_ftoa(long double n, char* res, int afterpoint);
 char* s21_uitoa(unsigned long int value, char* buffer, int base);
 char* repeat(char sym, int count);
 char* append(char* dest, char* append);
